Eliminate some compiler warnings.
* find/find.c: Remove definition of SAFE_CHDIR, which we don't
use.
* find/defs.h: Remove redundant declaration of launch.
* find/parser.c (parse_false): Cast unused arguments to void.
(parse_print0): Likewise.
(pred_context): Likewise.
(parse_newerXY): Add some parens for slightly greater clarity.
(make_segment): Avoid switch-missing-default-case warning by
turning it into an if statement.
(check_path_safety): Remove unused argument.
(insert_exec_ok): Don't pass the unwanted arugment to
check_path_safety.
(get_relative_timestamp): silence compiler warning by adding a
case for the remaining enumberation value rather than using
default.
* find/pred.c (months): the strings can be const char*, rather
than just char*.
(ctime_format): change TIME_BUF_LEN to an integer constant to
avoid signed/unsigned comparison.
(blank_rtrim): Change to new-style function definition(!) and
remove unnecessary parentheses around a return value.
* lib/buildcmd.c: Omit redundant declaration of environ.
* find/tree.c (get_expr): Make static.
(cost_assoc): make the name field const.
(prec_assoc): make the prec_name field const.
(op_assoc): make the type_name field const.
(type_name): turn into an ANSI function definition(!).
(prec_name): Likewise!  Also remove spurious parentheses around
return value.
(prec_name): Remove spurious parentheses around return value.
* lib/buildcmd.h (buildcmd_state): change types of several fields
to size_t: cmd_argc, cmd_argv_alloc, largest_successful_arg_count,
smallest_failed_arg_count.
(buildcmd_control): change types of several fields
to size_t: max_arg_count, initial_argc, lines_per_exec,
args_per_exec.

+++ lib/buildcmd.h
@@ -25,13 +25,13 @@
 struct buildcmd_state
 {
   /* Number of valid elements in `cmd_argv', including terminating NULL.  */
-  int cmd_argc;			/* 0 */
+  size_t cmd_argc;			/* 0 */
 
   /* The list of args being built.  */
   char **cmd_argv; /* NULL */
 
   /* Number of elements allocated for `cmd_argv'.  */
-  int cmd_argv_alloc;
+  size_t cmd_argv_alloc;
 
   /* Storage for elements of `cmd_argv'.  */
   char *argbuf;
@@ -52,8 +52,8 @@ struct buildcmd_state
   int dir_fd;
 
   /* Summary of what we think the argv limits are. */
-  int largest_successful_arg_count;
-  int smallest_failed_arg_count;
+  size_t largest_successful_arg_count;
+  size_t smallest_failed_arg_count;
 };
 
 struct buildcmd_control
@@ -77,7 +77,7 @@ struct buildcmd_control
    * function returns a useful value even if ARG_MAX is not defined.
    * However, sometimes, max_arg_count is LONG_MAX!
    */
-  long max_arg_count;
+  size_t max_arg_count;
 
 
   /* The length of `replace_pat'.  */
@@ -90,17 +90,17 @@ struct buildcmd_control
   char *replace_pat;
 
   /* Number of initial arguments given on the command line.  */
-  int initial_argc;		/* 0 */
+  size_t initial_argc;		/* 0 */
 
   /* exec callback. */
   int (*exec_callback)(struct buildcmd_control *, void *usercontext, int argc, char **argv);
 
   /* If nonzero, the maximum number of nonblank lines from stdin to use
      per command line.  */
-  long lines_per_exec;		/* 0 */
+  size_t lines_per_exec;		/* 0 */
 
   /* The maximum number of arguments to use per command line.  */
-  long args_per_exec;
+  size_t args_per_exec;
 };
 
 enum BC_INIT_STATUS
